\name{expResults-package}
\alias{expResults-package}
\docType{package}
\title{expResults}
%\description{}
\details{
\tabular{ll}{
Package: \tab expResults\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-06-28\cr
License: \tab GPL-2\cr
}
}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}

